chore (Setup): ESlint configuration - add Eslint configuration to enforce adherence to best practices and the approved coding convention. [Finishes #163426613]
