+++ .eslintignore
@@ -1,0 +1,2 @@
+# Dependency directory
+node_modules
