+++ .hound.yml
@@ -1,4 +1,5 @@
 eslint:
   enabled: true
-  config_file: .eslintrc.json
+  config_file: .eslintrc
   ignore_file: .eslintignore
+
