+++ package.json
@@ -3,10 +3,27 @@
   "version": "1.0.0",
   "description": "A Social platform for the creative at heart",
   "main": "index.js",
+  "engines": {
+    "node": ">=6.0.0"
+  },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 0",
-    "start": "node index.js",
-    "coverage": "npm test && nyc report --reporter=text-lcov | coveralls"
+    "start": "node server/index.js",
+    "test": "exit 0",
+    "coverage": "npm test && nyc report --reporter=text-lcov | coveralls",
+    "lint": "eslint '**/*js'",
+    "lint:fix": "prettier-eslint **/*{js,json} --write"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged",
+      "pre-push": "npm run test"
+    }
+  },
+  "lint-staged": {
+    "*.{js,json}": [
+      "prettier --write",
+      "git add"
+    ]
   },
   "author": "Andela Simulations Programme",
   "license": "MIT",
@@ -35,6 +52,11 @@
   "devDependencies": {
     "coveralls": "^3.0.2",
     "mocha-lcov-reporter": "^1.3.0",
-    "nyc": "^13.1.0"
+    "nyc": "^13.1.0",
+    "husky": "^1.3.1",
+    "lint-staged": "^8.1.1",
+    "prettier-cli": "^0.1.0",
+    "prettier-eslint": "^8.8.2",
+    "prettier-eslint-cli": "^4.7.1"
   }
 }
