+++ routes/api/index.js
@@ -2,17 +2,17 @@ const router = require("express").Router();
 
 router.use("/", require("./users"));
 
-router.use(function(err, req, res, next) {
-    if (err.name === "ValidationError") {
-        return res.status(422).json({
-            errors: Object.keys(err.errors).reduce(function(errors, key) {
-                errors[key] = err.errors[key].message;
-                return errors;
-            }, {})
-        });
-    }
+router.use((err, req, res, next) => {
+  if (err.name === "ValidationError") {
+    return res.status(422).json({
+      errors: Object.keys(err.errors).reduce((errors, key) => {
+        errors[key] = err.errors[key].message;
+        return errors;
+      }, {})
+    });
+  }
 
-    return next(err);
+  return next(err);
 });
 
 module.exports = router;
